fix: widen share distribution multiply to u128 to prevent latent overflow
The product `distributable * weights[i]` fits in u64 under current Zcash
supply assumptions (~2^27 * 2^32 = 2^59), but no code path enforces that
cap. Widening to u128 closes the latent overflow defensively.

+++ voting-circuits/src/vote_proof/builder.rs
@@ -175,7 +175,7 @@ fn distribute_remainder(
     // Integer division truncates, so we track how much was actually assigned.
     let mut assigned: u64 = 0;
     for i in 0..slots.len() {
-        let share = distributable * weights[i] / total_weight;
+        let share = ((distributable as u128 * weights[i] as u128) / total_weight as u128) as u64;
         slots[i] = 1 + share;
         assigned += share;
     }
